DBconfig movement
