+++ utils_base_env/__init__.py
@@ -5,4 +5,5 @@ from .utils_base_env import (
     str_arg_to_bool,
     add_sql_arguments,
     prepare_parser,
+    DbConfig
 )
